lib/ptabler/software/src/requirements.txt
polars-lts-cpu==1.30.0
polars-hash==0.5.4
polars-ds==0.9.0
msgspec==0.18.6
PyYAML==6.0.2
